+++ src/main/java/com/xiliulou/electricity/entity/payparams/WechatPublicKeyEntity.java
@@ -25,7 +25,7 @@ public class WechatPublicKeyEntity implements Serializable {
     private Long id;
     
     // 租户id
-    private Long tenantId;
+    private Integer tenantId;
     
     // 公钥ID
     private String pubKeyId;
