+++ src/main/java/com/xiliulou/electricity/vo/WechatPublicKeyVO.java
@@ -27,7 +27,7 @@ public class WechatPublicKeyVO implements Serializable {
     private Long id;
     
     // 租户id
-    private Long tenantId;
+    private Integer tenantId;
     
     // 公钥ID
     private String pubKeyId;
